+++ src/advanced/components/ui/CouponTab.tsx
@@ -2,7 +2,7 @@ import { CouponAddButton } from './CouponAddButton';
 import { CouponForm } from './CouponForm';
 import { CouponList } from './CouponList';
 import type { Coupon } from '../../../types';
-import type { NotificationVariant } from '../../constants';
+import type { NotificationVariant } from '../../entities/notification';
 import { useCouponForm } from '../../hooks/useCouponForm';
 
 interface CouponTabProps {
