+++ src/advanced/constants/index.ts
@@ -1,3 +1,2 @@
 export * from './coupon';
 export * from './product';
-export * from './notification';
