+++ src/advanced/entities/notification/consts/index.ts
