+++ src/advanced/hooks/useProductService.ts
@@ -1,7 +1,8 @@
 import { useCallback } from 'react';
 
-import type { NotificationVariant, ProductWithUI } from '../constants';
+import type { ProductWithUI } from '../constants';
 import { useProductStore } from './useProductStore';
+import type { NotificationVariant } from '../entities/notification';
 
 interface useProductServiceProps {
   onAddNotification: (message: string, type: NotificationVariant) => void;
