+++ src/advanced/pages/cart/ui/CartPage.tsx
@@ -3,7 +3,8 @@ import { CartItemList } from '../../../components/ui/CartItemList';
 import { CouponSelector } from '../../../components/ui/CouponSelector';
 import { PaymentSummary } from '../../../components/ui/PaymentSummary';
 import { ProductList } from '../../../components/ui/ProductList';
-import type { NotificationVariant, ProductWithUI } from '../../../constants';
+import type { ProductWithUI } from '../../../constants';
+import type { NotificationVariant } from '../../../entities/notification';
 import { useCartService } from '../../../hooks/useCartService';
 import { useDebouncedSearch } from '../../../hooks/useDebouncedSearch';
 import { calculateCartTotal } from '../../../models/cart';
